src/drfapp/requirements.txt
Django==2.2.13
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.1.0
djoser==2.0.3
Pillow==8.1.0

